src/test/javascript/spec/app/entities/sequence/sequence-update.component.spec.ts
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { FormBuilder } from '@angular/forms';
import { of } from 'rxjs';

import { LeloHubTestModule } from '../../../test.module';
import { SequenceUpdateComponent } from 'app/entities/sequence/sequence-update.component';
import { SequenceService } from 'app/entities/sequence/sequence.service';
import { Sequence } from 'app/shared/model/sequence.model';

describe('Component Tests', () => {
  describe('Sequence Management Update Component', () => {
    let comp: SequenceUpdateComponent;
    let fixture: ComponentFixture<SequenceUpdateComponent>;
    let service: SequenceService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [LeloHubTestModule],
        declarations: [SequenceUpdateComponent],
        providers: [FormBuilder]
      })
        .overrideTemplate(SequenceUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(SequenceUpdateComponent);
      comp = fixture.componentInstance;
      service = fixture.debugElement.injector.get(SequenceService);
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', fakeAsync(() => {
        // GIVEN
        const entity = new Sequence(123);
        spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.update).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));

      it('Should call create service on save for new entity', fakeAsync(() => {
        // GIVEN
        const entity = new Sequence();
        spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.create).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));
    });
  });
});
